requirements.txt
absl-py==0.14.1
argcomplete==1.12.3
argon2-cffi==21.1.0
astunparse==1.6.3
attrs==21.2.0
backcall==0.2.0
bleach==4.1.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.7
colorama==0.4.4
cycler==0.10.0
debugpy==1.5.0
decorator==5.1.0
defusedxml==0.7.1
easydict==1.9
entrypoints==0.3
gast==0.3.3
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.41.0
h5py==2.10.0
idna==3.2
importlib-metadata==4.8.1
ipykernel==6.5.0
ipython==7.29.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.2
jsonschema==4.2.1
jupyter-client==7.0.6
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.4
numpy==1.18.5
oauthlib==3.1.1
opencv-python==4.5.3.56
opt-einsum==3.3.0
packaging==21.0
pandocfilters==1.5.0
parso==0.8.2
pickleshare==0.7.5
Pillow==8.3.2
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.18.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pywin32==302
pywinpty==1.1.4
pyzmq==22.3.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.4.1
Send2Trash==1.8.0
six==1.16.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
tornado==6.1
traitlets==5.1.0
typing-extensions==3.10.0.2
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
wrapt==1.13.1
zipp==3.6.0
